separate demo view
