+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/dev/WorkbookDemoApp.kt
@@ -15,7 +15,7 @@ class WorkbookDemoApp : App(RootView::class) {
     }
 
     override fun onBeforeShow(view: UIComponent) {
-        workspace.dock<WorkbookTableDemoView>() // set the view for demo here
+        workspace.dock<TranslationCardsDemoView>() // set the view for demo here
         workspace.root.apply {
             contextmenu {
                 item("Change Theme") {
