+++ jvm/workbookapp/src/main/kotlin/org/wycliffeassociates/otter/jvm/workbookapp/ui/dev/WorkbookTableDemoView.kt
@@ -1,19 +1,11 @@
 package org.wycliffeassociates.otter.jvm.workbookapp.ui.dev
 
-import javafx.beans.property.SimpleBooleanProperty
-import javafx.beans.property.SimpleObjectProperty
-import org.wycliffeassociates.otter.common.data.primitives.Language
 import org.wycliffeassociates.otter.common.data.workbook.WorkbookInfo
-import org.wycliffeassociates.otter.jvm.controls.card.CreateTranslationCard
-import org.wycliffeassociates.otter.jvm.controls.card.activeTranslationCard
-import org.wycliffeassociates.otter.jvm.controls.card.newTranslationCard
-import org.wycliffeassociates.otter.jvm.controls.card.translationCard
-import org.wycliffeassociates.otter.jvm.controls.card.translationCardWrapper
 import org.wycliffeassociates.otter.jvm.controls.event.WorkbookDeleteEvent
 import org.wycliffeassociates.otter.jvm.controls.event.WorkbookExportEvent
 import org.wycliffeassociates.otter.jvm.controls.event.WorkbookOpenEvent
-import org.wycliffeassociates.otter.jvm.controls.model.TranslationMode
 import org.wycliffeassociates.otter.jvm.controls.styles.tryImportStylesheet
+import org.wycliffeassociates.otter.jvm.workbookapp.ui.components.tableview.workbookTableView
 import java.time.LocalDateTime
 import tornadofx.*
 
@@ -31,21 +23,13 @@ class WorkbookTableDemoView : View() {
         WorkbookInfo(0, "Colossians", "", 1.0, LocalDateTime.now(), true),
     )
 
-    val languages = listOf(
-        Language("en", "English", "English", "", true, ""),
-        Language("fr", "français", "French", "", true, ""),
-    )
-
     init {
         tryImportStylesheet("/css/popup-menu.css")
         tryImportStylesheet("/css/filtered-search-bar.css")
         tryImportStylesheet("/css/table-view.css")
-        tryImportStylesheet("/css/translation-card-2.css")
 
         subscribeToWorkbookEvent()
     }
-
-    private val showNewTranslationCard = SimpleBooleanProperty(false)
 
     private fun subscribeToWorkbookEvent() {
         workspace.subscribe<WorkbookOpenEvent> {
@@ -63,48 +47,7 @@ class WorkbookTableDemoView : View() {
     }
 
     override val root = vbox {
-        spacing = 10.0
         paddingAll = 20.0
-        maxWidth = 300.0
-
-        borderpane {
-            center = translationCardWrapper(
-                languages[0],
-                languages[1],
-                TranslationMode.TRANSLATION
-            ) {
-                top = button("Reset") {
-                    action {
-                        this@translationCardWrapper.isActiveProperty.set(false)
-                        showNewTranslationCard.set(false)
-                    }
-                }
-            }
-        }
-
-        newTranslationCard(
-            SimpleObjectProperty<Language>(
-                Language("en", "English", "English", "", true, "")
-            ),
-            SimpleObjectProperty<Language>(null),
-            mode = TranslationMode.NARRATION
-        ) {
-            visibleWhen(showNewTranslationCard)
-            managedWhen(visibleProperty())
-
-            setOnCancelAction {
-                showNewTranslationCard.set(false)
-            }
-        }
-        add(
-            CreateTranslationCard().apply {
-                visibleWhen(showNewTranslationCard.not())
-                managedWhen(visibleProperty())
-
-                setOnAction {
-                    showNewTranslationCard.set(true)
-                }
-            }
-        )
+        workbookTableView(workbookList)
     }
 }
